+++ .atomist/tests/handlers/command/TypeScriptCommandHandlerSteps.ts
@@ -1,0 +1,16 @@
+import { Given, When, Then, HandlerScenarioWorld, CommandHandlerScenarioWorld } from "@atomist/rug/test/handler/Core"
+
+Given("nothing", f => { });
+
+When("the TypeScriptCommandHandler is invoked", (world: HandlerScenarioWorld) => {
+    let w: CommandHandlerScenarioWorld = world as CommandHandlerScenarioWorld;
+    let handler = w.commandHandler("TypeScriptCommandHandler");
+    w.invokeHandler(handler, {});
+});
+
+Then("you get the right response", (world: HandlerScenarioWorld) => {
+    let w: CommandHandlerScenarioWorld = world as CommandHandlerScenarioWorld;
+    const expected = "Successfully ran TypeScriptCommandHandler: default value";
+    const message = w.plan().messages[0].body.value;
+    return message == expected;
+});